Fix errors, organize packages, reorder calls
