+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".presentation.view.activity.SplashScreenActivity"
+            android:name=".home.presentation.view.activity.SplashScreenActivity"
             android:label="@string/app_name"
             android:theme="@style/splash_theme">
             <intent-filter>
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.view.activity.HomeScreenActivity"
+            android:name=".home.presentation.view.activity.HomeScreenActivity"
             android:label="@string/title_activity_home_screen" />
     </application>
