--- app/src/main/java/com/globant/example/mentorapp/data/util/ApiUtils.java
@@ -1,16 +1,0 @@
-package com.globant.example.mentorapp.data.util;
-
-/**
- * Class to define some communications tools.
- * Created by oscar.amezquita on 7/06/2017.
- */
-
-public class ApiUtils {
-
-    public static final String BASE_URL = "https://api.github.com";
-    public static final String PROFILE_NAME_GITHUB = "googlesamples";
-    public static final String PROFILE_REPOSITORY = "android-architecture";
-    public static final String SERVICE_RESPONSE_OK = "list ok";
-    public static final String SERVICE_RESPONSE_COMM_ERROR = "COMM error";
-
-}
