+++ app/src/main/java/com/globant/example/mentorapp/home/domain/interactor/ListUsersInteractor.java
@@ -1,0 +1,13 @@
+package com.globant.example.mentorapp.home.domain.interactor;
+
+/**
+ * Created by oscar.amezquita on 9/06/2017.
+ */
+
+public interface ListUsersInteractor {
+
+    /**
+     * Find a list of users from a provider
+     */
+    void getUsersList();
+}
