+++ app/src/main/java/com/globant/example/mentorapp/home/domain/model/UserEntity.java
@@ -1,0 +1,173 @@
+package com.globant.example.mentorapp.home.domain.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class UserEntity {
+
+    @Expose
+    private String login;
+    @Expose
+    private Integer id;
+    @SerializedName("avatar_url")
+    @Expose
+    private String avatarUrl;
+    @SerializedName("gravatar_id")
+    @Expose
+    private String gravatarId;
+    @Expose
+    private String url;
+    @SerializedName("html_url")
+    @Expose
+    private String htmlUrl;
+    @SerializedName("followers_url")
+    @Expose
+    private String followersUrl;
+    @SerializedName("following_url")
+    @Expose
+    private String followingUrl;
+    @SerializedName("gists_url")
+    @Expose
+    private String gistsUrl;
+    @SerializedName("starred_url")
+    @Expose
+    private String starredUrl;
+    @SerializedName("subscriptions_url")
+    @Expose
+    private String subscriptionsUrl;
+    @SerializedName("organizations_url")
+    @Expose
+    private String organizationsUrl;
+    @SerializedName("repos_url")
+    @Expose
+    private String reposUrl;
+    @SerializedName("events_url")
+    @Expose
+    private String eventsUrl;
+    @SerializedName("received_events_url")
+    @Expose
+    private String receivedEventsUrl;
+    @Expose
+    private String type;
+    @SerializedName("site_admin")
+    @Expose
+    private Boolean siteAdmin;
+
+    /**
+     * Entity to save Users general information
+     *
+     */
+    public UserEntity() {
+    }
+
+    /**
+     *
+     * @param eventsUrl
+     * @param siteAdmin
+     * @param gistsUrl
+     * @param type
+     * @param gravatarId
+     * @param url
+     * @param subscriptionsUrl
+     * @param id
+     * @param followersUrl
+     * @param reposUrl
+     * @param htmlUrl
+     * @param receivedEventsUrl
+     * @param avatarUrl
+     * @param followingUrl
+     * @param login
+     * @param organizationsUrl
+     * @param starredUrl
+     */
+    public UserEntity(String login, Integer id, String avatarUrl, String gravatarId, String url, String htmlUrl, String followersUrl, String followingUrl, String gistsUrl, String starredUrl, String subscriptionsUrl, String organizationsUrl, String reposUrl, String eventsUrl, String receivedEventsUrl, String type, Boolean siteAdmin) {
+        super();
+        this.login = login;
+        this.id = id;
+        this.avatarUrl = avatarUrl;
+        this.gravatarId = gravatarId;
+        this.url = url;
+        this.htmlUrl = htmlUrl;
+        this.followersUrl = followersUrl;
+        this.followingUrl = followingUrl;
+        this.gistsUrl = gistsUrl;
+        this.starredUrl = starredUrl;
+        this.subscriptionsUrl = subscriptionsUrl;
+        this.organizationsUrl = organizationsUrl;
+        this.reposUrl = reposUrl;
+        this.eventsUrl = eventsUrl;
+        this.receivedEventsUrl = receivedEventsUrl;
+        this.type = type;
+        this.siteAdmin = siteAdmin;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public String getAvatarUrl() {
+        return avatarUrl;
+    }
+
+    public String getGravatarId() {
+        return gravatarId;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public String getHtmlUrl() {
+        return htmlUrl;
+    }
+
+    public String getFollowersUrl() {
+        return followersUrl;
+    }
+
+    public String getFollowingUrl() {
+        return followingUrl;
+    }
+
+    public String getGistsUrl() {
+        return gistsUrl;
+    }
+
+    public String getStarredUrl() {
+        return starredUrl;
+    }
+
+    public String getSubscriptionsUrl() {
+        return subscriptionsUrl;
+    }
+
+    public String getOrganizationsUrl() {
+        return organizationsUrl;
+    }
+
+    public String getReposUrl() {
+        return reposUrl;
+    }
+
+    public String getEventsUrl() {
+        return eventsUrl;
+    }
+
+
+    public String getReceivedEventsUrl() {
+        return receivedEventsUrl;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public Boolean getSiteAdmin() {
+        return siteAdmin;
+    }
+
+}
