+++ app/src/main/java/com/globant/example/mentorapp/home/presentation/model/ModelUserEntity.java
@@ -1,0 +1,30 @@
+package com.globant.example.mentorapp.home.presentation.model;
+
+/**
+ * Created by oscar.amezquita on 16/06/2017.
+ */
+
+public class ModelUserEntity {
+
+    private String name;
+    private String imageUrl;
+    private String profileUrl;
+
+    public ModelUserEntity(String name, String imageUrl, String profileUrl) {
+        this.name = name;
+        this.imageUrl = imageUrl;
+        this.profileUrl = profileUrl;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public String getProfileUrl() {
+        return profileUrl;
+    }
+}
