--- app/src/main/java/com/globant/example/mentorapp/presentation/view/fragment/ListUsersInterface.java
@@ -1,18 +1,0 @@
-package com.globant.example.mentorapp.presentation.view.fragment;
-
-/**
- * Interface to define List of users  basic behaviors
- * Created by oscar.amezquita on 9/06/2017.
- */
-
-public interface ListUsersInterface {
-
-    void usersReady();
-
-    void usersErrorHttp();
-
-    void usersErrorConnectivity();
-
-    void snackBarMessage(int stringResource);
-
-}
