+++ app/src/main/res/layout/activity_home_screen.xml
@@ -3,8 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.activity.HomeScreenActivity">
-
+    tools:context=".home.presentation.view.activity.HomeScreenActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
