+++ app/src/main/res/layout/fragment_home_screen.xml
@@ -3,12 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.fragment.ListUsersFragment"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    tools:context=".home.presentation.view.fragment.ListUsersFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/list_item_container"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent"
+        tools:listitem="@layout/list_users_adapter" />
 
 </LinearLayout>
