+++ app/src/main/res/layout/list_users_adapter.xml
@@ -14,6 +14,7 @@
             android:id="@+id/imageUser"
             android:layout_width="@dimen/list_user_image_width"
             android:layout_height="@dimen/list_user_image_height"
+            android:adjustViewBounds="true"
             android:paddingTop="@dimen/list_general_margin"
             app:srcCompat="@mipmap/ic_launcher" />
 
@@ -22,7 +23,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_margin="@dimen/list_general_margin"
-            android:textAppearance="@style/TextAppearance.AppCompat.Body1"
+            android:textAppearance="@style/TextAppearance.AppCompat.Display1"
+            android:textSize="24sp"
             tools:text="UserName" />
     </LinearLayout>
