+++ app/src/main/res/menu/menu_home_screen.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.view.activity.HomeScreenActivity">
+    tools:context=".home.presentation.view.activity.HomeScreenActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
